+++ dist-esnext/__testdata__/index.data.test.js
@@ -1,0 +1,3 @@
+export * from './board.data.test';
+export * from './game.data.test';
+//# sourceMappingURL=index.data.test.js.map