+++ src/__testdata__/board.data.test.ts
@@ -1,5 +1,27 @@
 import { Board, Position } from '../index';
 
+const startBlackPiecesExpected = [
+  { x: 0, y: 0, isBlack: true },
+  { x: 1, y: 0, isBlack: true },
+  { x: 2, y: 0, isBlack: true },
+  { x: 3, y: 0, isBlack: true },
+  { x: 4, y: 0, isBlack: true },
+  { x: 5, y: 0, isBlack: true },
+  { x: 6, y: 0, isBlack: true },
+  { x: 7, y: 0, isBlack: true }
+];
+
+const startWhitePiecesExpected = [
+  { x: 0, y: 7, isBlack: false },
+  { x: 1, y: 7, isBlack: false },
+  { x: 2, y: 7, isBlack: false },
+  { x: 3, y: 7, isBlack: false },
+  { x: 4, y: 7, isBlack: false },
+  { x: 5, y: 7, isBlack: false },
+  { x: 6, y: 7, isBlack: false },
+  { x: 7, y: 7, isBlack: false }
+];
+
 const startPiecesExpected = [
   { x: 7, y: 0, isBlack: true }, { x: 7, y: 7, isBlack: false },
   { x: 6, y: 0, isBlack: true }, { x: 6, y: 7, isBlack: false },
@@ -78,7 +100,9 @@ export {
   cleanBoardExpected,
   initialBoard,
   initialBoardExpected,
+  startBlackPiecesExpected,
   startPiecesExpected,
+  startWhitePiecesExpected,
   unicodeStartBoard,
   xAndYStartBoard
 };