+++ src/__testdata__/index.data.test.ts
@@ -1,0 +1,2 @@
+export * from './board.data.test';
+export * from './game.data.test';